requirements.txt
Flask==2.3.3
graphviz==0.20.1
networkx==3.1
openai==0.28.0
beautifulsoup4==4.12.2
neo4j==5.12.0
python-dotenv==1.0.0